% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanr.R
\name{cleanr}
\alias{cleanr}
\title{Basic data cleaning}
\usage{
cleanr(data, x, n)
}
\arguments{
\item{data}{the dataset to be cleaned}

\item{x}{the numeric vector to be cleaned}

\item{n}{the threshold used to clean data}
}
\value{
returns a cleaned dataset
}
\description{
Basic data cleaning
}
\examples{
x<-c(100,100,100,100,-100,-100)
y<-c(100,100,100,100,100,100)
mydata<-data.frame(x,y)
cleanr(mydata,x,0)

}
